Install latest puppet-vcsrepo
Puppet4 is a lot pickier about puppet module dependency resolution. On
some hosts with puppet4 we are getting:

  ModuleLoader: module 'project_config' has unresolved dependencies - it will only see those that are resolved. Use 'puppet module list --tree' to see information about modules

Running the suggested command we see that vcsrepo is not resolved
because the version of vcsrepo we install doesn't ahve a metadata.json.
Thankfully the HEAD of vcsrepo does have metadata.json which should
allow this to be resolved.

Depends-On: https://review.openstack.org/641161
Change-Id: I022afd8bab58588e31f3f8e806183933a1de4e6b