Only install puppet 4 if not already installed
The reinstallation of puppet 4 on every ansible run was assumed to be a
non issue that we could disable after our complete conversion to puppet
4 (via removal of the explicit install playbook and reliance on our
launch tooling). However, Installation of this package actually causes
the removal of Gerrit rc.d files which the next puppet run reinstalls.

Avoid this back and forth and the unnecesasry additional package
installation by checking if the puppet 4 bin path exists prior to
reinstalling puppet 4.

Change-Id: I47821ae799412f6cb9693b617ae6e7e630590c15