Add global inventory to launch_node
Passing the -i to the jobdir means we're overriding the inventory.
This means variables that come from the /etc/ansible vars, like
sysadmins, are missing.

Add the global inventory to the command line for ansible-playbook.
We have --limit specified from '-l' - so we should still only run
on the host in question.

Change-Id: Ia67e65d25a1d961b619aa445303015fd577dee57