Add a soft dependency between gitea build and run
When we run the run-gitea job, make sure that it runs *after*
the gitea image build job, if the build job ran at all.  If
it didn't, we don't need to worry about it.

This has to happen in the project pipeline config because the
job dependency is different in check and gate.

Change-Id: I7cb069a6cd40a4ae8d5cbcdf23e7686b301493a1