Use opendev logos
The stacked opendev logo is added to the home page as an svg.

A 120x120 ping is added as the gitea icon.  It appears in several
places throughout gitea source code, and therefore it's easier to
just replace it using the same filename and format rather than
updating the templates with a new filename.

Change-Id: I888b8913c7b6a1045975f10240d744d09b9355b0