Set the gitea theme color to match the opendev pink
Currently the default is to use gitea green which looks a little weird
on our site when using mobile browsers. I don't see an easy way to
revert to the browsers default color but there may be some magic like
using a 'default' string? That may be an alternative option we can
consider.

Change-Id: Ia4e3d25b75bba169c3b7cc60c52c0de791e6be21