Explicitly set up mirror update crons under root user
We were using $user as the user to setup the crons under which looked up
the user fact for the user running puppet. This meant that we got root.
When we upgraded to puppet 4 the fact is actually $::user, but there
isn't really a good reason to make this variable. If we start running
pupept under a different user we'll get multiple crons configured and it
will be a mess.

Explicitly set up the crons under the root user for clarity and
simplicity.

Change-Id: If7841bf95d252c1fd358bf329e6ea71f58e14a30