Remove /hosts from log outputs
Change I754637115f8c7469efbc1856e88bbcb6fb83b4ce moved a bunch of log
collection to use "stage-output".  This uses "fetch-output" which
automatically puts these logs in hostname subdirectories; but it does
not have an option to put it in hosts/hostname as we were doing with
the other logs.

Although we could add such support, it probably doesn't make sense as
most other multinode jobs will have the same layout with the host logs
at the top level.  Remove the intermediate "/hosts/" directory on
system-config jobs so all logs remain at the top level, and we don't
have this confusing split as to where logs are for each host.

Change-Id: I56bd67c659ffb26a460d9406f6f090d431c8aa79