Add support for LE roles to restart apache
As we transition from purchased certs managed by puppet to LE certs
managed via ansible and acme we need some way of telling apache to
restart when it has new certs. We do this via a flag that the LE cert
install role checks and will restart apache if set and new certs were
created for this host.

Note that apache must be restarted not reloaded to see the new certs.

Change-Id: I3c354ec2dffa19859c97b222cbdf030bd2363bac