Pivot to running zuul cd against nameservers
Instead of trying to zuul cd the zuul scheduler project updates which
are partially managed by puppet we pivot to managing the nameservers in
this way. This is a better candidate because they are managed entirely
with ansible.

To start we just check that we can successfully run a hello world
playbook against the nameservers and continue to actually manage them
via base.yaml. Once we have the simple playbook working we can update
thinsg to move away from the base.yaml cron loop. When we do this we
will want to add this job to our zonefile repos as well.

Depends-On: https://review.openstack.org/651389
Change-Id: I77252a7680e9bfb50efe19e1f4241ccda239e50b