Update grafana for new archive repo
Grafana has moved to a new package repo [1] and the apt-get update
step is failing on the current host.

The first version of puppet-grafana that has this update is 6.0.0;
this is two years later than the current version we're using so
... yeah.

It looks like upstream has moved from camptocamp/archive to
voxpupuli/archive so the comment is no longer required.

[1] https://grafana.com/blog/2019/01/05/moving-to-packages.grafana.com/

Change-Id: Ibab0ed6799563ba0f9674ef1ea575c6ac5d60341