[wip] letsencrypt update idea
the idea is that hosts can define a handler that listens for an event
triggered when their certificates are updated.  they can then restart
services, etc.

I'm not that confident this works; there seems to be an active ansible
bug [1] and not sure we can template out listener/handler signals ...

[1] https://github.com/ansible/ansible/issues/47287

Change-Id: I4e0609c4751643d6e0c8d9eaa38f184e0ce5452e