Add a opendev migration repo name mapping script
Git repo moves based on cgit aliases from project-config, the
OpenStack TC guidance recorded in
http://lists.openstack.org/pipermail/openstack-discuss/2019-April/004920.html
and the ethercalc used to collect input from other users of the
system. Also the results of an extensive bikeshedding session at
http://eavesdrop.openstack.org/irclogs/%23openstack-infra/%23openstack-infra.2019-04-11.log.html#t2019-04-11T14:54:09
which concluded that anything left homeless goes in a namespace
called "x" since that's short, a basic alphabetic character and
provides no particular connotation.

This script, when run, provides a shareable rendering on stdout and
also writes a repos.yaml file for input into the rename_repos
playbook.

Change-Id: Ifa9fa6896110e8a33f32dcda6325bd58846935e2